requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pandas==2.2.0
python-dotenv==1.0.1
llama-index==0.10.1
llama-index-llms-huggingface==0.1.4
llama-index-embeddings-huggingface==0.1.4
sentence-transformers==2.3.1
faiss-cpu==1.7.4
transformers==4.37.2
torch==2.2.0 